migrate the comparing of the freshest data methodologies into tests and the filtering into the transform step
